+++ test/resource.test.ts
@@ -16,12 +16,16 @@
 
 import { strict as assert } from 'assert';
 import { beforeEach, describe, it } from 'node:test';
-import { detect } from '../src/resource';
-import * as utils from './utils';
+import { cleanEnvironment, detectResource } from './utils';
+import {
+  ATTR_TELEMETRY_SDK_LANGUAGE,
+  ATTR_TELEMETRY_SDK_VERSION,
+  ATTR_TELEMETRY_SDK_NAME,
+} from '@opentelemetry/semantic-conventions';
 
 describe('resource detector', () => {
   beforeEach(() => {
-    utils.cleanEnvironment();
+    cleanEnvironment();
   });
 
   describe('resource.detect', () => {
@@ -29,22 +33,22 @@ describe('resource detector', () => {
       process.env.OTEL_RESOURCE_ATTRIBUTES =
         'k=v,service.name=node-svc,x=a%20b';
 
-      const resource = detect();
-      assert.strictEqual(resource.attributes['k'], 'v');
-      assert.strictEqual(resource.attributes['x'], 'a b');
-      assert.strictEqual(resource.attributes['service.name'], 'node-svc');
+      const resource = detectResource();
+      assert.strictEqual(resource.attributes?.['k'], 'v');
+      assert.strictEqual(resource.attributes?.['x'], 'a b');
+      assert.strictEqual(resource.attributes?.['service.name'], 'node-svc');
     });
 
     it('catches service name from the env', () => {
       process.env.OTEL_SERVICE_NAME = 'node-svc';
-      const resource = detect();
+      const resource = detectResource();
 
-      assert.strictEqual(resource.attributes['service.name'], 'node-svc');
+      assert.strictEqual(resource.attributes?.['service.name'], 'node-svc');
     });
 
     it('catches process attributes', () => {
-      const resource = detect();
-      assert.strictEqual(typeof resource.attributes['process.pid'], 'number');
+      const resource = detectResource();
+      assert.strictEqual(typeof resource.attributes?.['process.pid'], 'number');
 
       for (const attributeName of [
         'process.executable.name',
@@ -52,20 +56,40 @@ describe('resource detector', () => {
         'process.runtime.version',
         'process.runtime.name',
       ]) {
-        assert.strictEqual(typeof resource.attributes[attributeName], 'string');
+        assert.strictEqual(
+          typeof resource.attributes?.[attributeName],
+          'string'
+        );
       }
     });
 
     it('detects OS attributes', () => {
-      const resource = detect();
-      assert.strictEqual(typeof resource.attributes['os.type'], 'string');
-      assert.strictEqual(typeof resource.attributes['os.version'], 'string');
+      const resource = detectResource();
+      assert.strictEqual(typeof resource.attributes?.['os.type'], 'string');
+      assert.strictEqual(typeof resource.attributes?.['os.version'], 'string');
     });
 
     it('catches host attributes', () => {
-      const resource = detect();
-      assert.strictEqual(typeof resource.attributes['host.name'], 'string');
-      assert.strictEqual(typeof resource.attributes['host.arch'], 'string');
+      const resource = detectResource();
+      assert.strictEqual(typeof resource.attributes?.['host.name'], 'string');
+      assert.strictEqual(typeof resource.attributes?.['host.arch'], 'string');
+    });
+
+    it('adds telemetry.sdk attributes', () => {
+      const resource = detectResource();
+
+      assert.strictEqual(
+        typeof resource.attributes?.[ATTR_TELEMETRY_SDK_LANGUAGE],
+        'string'
+      );
+      assert.strictEqual(
+        typeof resource.attributes?.[ATTR_TELEMETRY_SDK_NAME],
+        'string'
+      );
+      assert.strictEqual(
+        typeof resource.attributes?.[ATTR_TELEMETRY_SDK_VERSION],
+        'string'
+      );
     });
   });
 });